+++ packages/babel-preset-expo/src/index.ts
@@ -204,16 +204,19 @@ function babelPresetExpo(api: ConfigAPI, options: BabelPresetExpoOptions = {}): 
     // getters and setters in spread objects. We need to add this plugin ourself without that option.
     // @see https://github.com/expo/expo/pull/11960#issuecomment-887796455
     extraPlugins.push([require('@babel/plugin-transform-object-rest-spread'), { loose: false }]);
-  } else if (!isServerEnv) {
-    // This is added back on hermes to ensure the react-jsx-dev plugin (`@babel/preset-react`) works as expected when
-    // JSX is used in a function body. This is technically not required in production, but we
-    // should retain the same behavior since it's hard to debug the differences.
-    extraPlugins.push(require('@babel/plugin-transform-parameters'));
+  } else {
+    if (platform !== 'web' && !isServerEnv) {
+      // This is added back on hermes to ensure the react-jsx-dev plugin (`@babel/preset-react`) works as expected when
+      // JSX is used in a function body. This is technically not required in production, but we
+      // should retain the same behavior since it's hard to debug the differences.
+      extraPlugins.push(require('@babel/plugin-transform-parameters'));
+    }
   }
 
-  const inlines: Record<string, boolean | string> = {
+  const inlines: Record<string, null | boolean | string> = {
     'process.env.EXPO_OS': platform,
     // 'typeof document': isServerEnv ? 'undefined' : 'object',
+    'process.env.EXPO_SERVER': !!isServerEnv,
   };
 
   // `typeof window` is left in place for native + client environments.