+++ packages/expo-audio/android/src/main/java/expo/modules/audio/AudioRecords.kt
@@ -13,14 +13,13 @@ class AudioSource(
 
 class AudioMode(
   @Field val playsInSilentMode: Boolean = false,
-  @Field val interruptionMode: InterruptionMode = InterruptionMode.MIX_WITH_OTHERS,
+  @Field val interruptionMode: InterruptionMode = InterruptionMode.DO_NOT_MIX,
   @Field val allowsRecording: Boolean = true,
   @Field val shouldPlayInBackground: Boolean = true,
   @Field val shouldRouteThroughEarpiece: Boolean?
 ) : Record
 
 enum class InterruptionMode(val value: String) : Enumerable {
-  MIX_WITH_OTHERS("mixWithOthers"),
   DO_NOT_MIX("doNotMix"),
   DUCK_OTHERS("duckOthers")
 }