+++ packages/expo-contacts/ios/ContactsRequester.swift
@@ -13,9 +13,15 @@ class ContactsPermissionRequester: NSObject, EXPermissionsRequester {
     switch permissions {
     case .authorized:
       status = EXPermissionStatusGranted
+    #if compiler(>=6)
+    case .limited:
+      status = EXPermissionStatusGranted
+    #endif
     case .denied, .restricted:
       status = EXPermissionStatusDenied
     case .notDetermined:
+      status = EXPermissionStatusUndetermined
+    @unknown default:
       status = EXPermissionStatusUndetermined
     }