+++ packages/expo-image-manipulator/ios/ImageRef.swift
@@ -1,0 +1,8 @@
+// Copyright 2024-present 650 Industries. All rights reserved.
+
+import ExpoModulesCore
+
+/**
+ Represents a shared reference to the `UIImage` instance.
+ */
+internal final class ImageRef: SharedRef<UIImage> {}